refactor: Move DocBook generation to separate file
